- refactoring for quilt as authid current user so it can be used as component - plsql_profiler data are stored in current user schema objects

+++ application/schema/package/quilt_reported_objects.pkb
@@ -1,31 +1,15 @@
 CREATE OR REPLACE PACKAGE BODY quilt_reported_objects AS
 
     ----------------------------------------------------------------------------  
-    PROCEDURE enable_report
-    (
-        p_owner       IN all_objects.owner%Type,
-        p_object_name IN all_objects.object_name%Type DEFAULT NULL
-    ) IS
+    PROCEDURE enable_report(p_objects IN quilt_object_list_type) IS
         PRAGMA AUTONOMOUS_TRANSACTION;
     BEGIN
-        quilt_logger.log_detail('begin:p_owner=$1,p_object_name=$2', p_owner, p_object_name);
+        quilt_logger.log_detail('begin');
         MERGE INTO quilt_reported_object t
-        USING (SELECT sys_context('USERENV', 'SESSIONID') AS sessionid,
-                      sys_context('USERENV', 'SID') AS SID,
-                      OWNER,
-                      object_name,
-                      object_type
-                 FROM all_objects
-                WHERE OWNER = p_owner
-                  AND object_name LIKE '%' || p_object_name || '%'
-                  AND object_type IN (quilt_const.OBJ_TYPE_PACKAGE_BODY,
-                                      quilt_const.OBJ_TYPE_TYPE_BODY,
-                                      quilt_const.OBJ_TYPE_PROCEDURE,
-                                      quilt_const.OBJ_TYPE_FUNCTION,
-                                      quilt_const.OBJ_TYPE_TRIGGER)) s
-        ON (s.owner = t.owner AND s.object_name = t.object_name AND s.object_type = t.object_type AND s.sessionid = t.sessionid AND s.sid = t.sid)
+        USING (SELECT * FROM TABLE(p_objects)) s
+        ON (s.owner = t.owner AND s.object_name = t.object_name AND s.object_type = t.object_type)
         WHEN NOT MATCHED THEN
-            INSERT (sessionid, SID, OWNER, object_name, object_type) VALUES (s.sessionid, s.sid, s.owner, s.object_name, s.object_type);
+            INSERT (OWNER, object_name, object_type) VALUES (s.owner, s.object_name, s.object_type);
         COMMIT;
         quilt_logger.log_detail('end');
     EXCEPTION
@@ -36,17 +20,11 @@ CREATE OR REPLACE PACKAGE BODY quilt_reported_objects AS
     END enable_report;
 
     ----------------------------------------------------------------------------  
-    PROCEDURE disable_report
-    (
-        p_owner       IN all_objects.owner%Type,
-        p_object_name IN all_objects.object_name%Type DEFAULT NULL
-    ) IS
+    PROCEDURE disable_report(p_objects IN quilt_object_list_type) IS
         PRAGMA AUTONOMOUS_TRANSACTION;
     BEGIN
-        quilt_logger.log_detail('begin:p_owner=$1,p_object_name=$2', p_owner, p_object_name);
-        DELETE FROM quilt_reported_object
-         WHERE OWNER = p_owner
-           AND object_name LIKE '%' || p_object_name || '%';
+        quilt_logger.log_detail('begin');
+        DELETE FROM quilt_reported_object WHERE (OWNER, object_name, object_type) IN (SELECT * FROM TABLE(p_objects));
         COMMIT;
         quilt_logger.log_detail('end');
     EXCEPTION
@@ -57,26 +35,42 @@ CREATE OR REPLACE PACKAGE BODY quilt_reported_objects AS
     END disable_report;
 
     ----------------------------------------------------------------------------
-    FUNCTION get_reported_objects
-    (
-        p_sessionId IN NUMBER DEFAULT NULL,
-        p_sid       IN NUMBER DEFAULT NULL
-    ) RETURN SYS_REFCURSOR IS
-        l_sessionId NUMBER := nvl(p_sessionId, quilt_core.get_sessionId);
-        l_SID       NUMBER := nvl(p_sid, quilt_core.get_SID);
-        lrcu_result SYS_REFCURSOR;
+    FUNCTION get_reported_objects RETURN quilt_object_list_type IS
+        ltab_result quilt_object_list_type;
     BEGIN
-        quilt_logger.log_detail('begin:p_sessionId=$1, p_SID=$2, l_sessionId=$3, l_SID=$4', p_sessionId, p_SID, l_sessionId, l_SID);
+        quilt_logger.log_detail('begin');
         --
-        OPEN lrcu_result FOR
-            SELECT OWNER, object_name, object_type
-              FROM quilt_reported_object
-             WHERE SID = l_SID
-               AND sessionid = l_sessionId;
+        SELECT quilt_object_type(OWNER, object_name, object_type) BULK COLLECT INTO ltab_result FROM quilt_reported_object;
         --
-        RETURN lrcu_result;
+        RETURN ltab_result;
         --
     END get_reported_objects;
+
+    ----------------------------------------------------------------------------
+    PROCEDURE save_source
+    (
+        p_quilt_run_id IN INTEGER,
+        p_sources      IN typ_source_tab
+    ) IS
+        PRAGMA AUTONOMOUS_TRANSACTION;
+    BEGIN
+        -- TODO: merge?
+        FORALL idx IN p_sources.first .. p_sources.last
+            INSERT INTO quilt_reported_object_source
+            VALUES
+                (p_quilt_run_id,
+                 p_sources(idx).owner,
+                 p_sources(idx).name,
+                 p_sources(idx).type,
+                 p_sources(idx).line,
+                 p_sources(idx).text,
+                 p_sources(idx).origin_con_id);
+        COMMIT;
+    EXCEPTION
+        WHEN OTHERS THEN
+            ROLLBACK;
+            RAISE;
+    END;
 
 END quilt_reported_objects;
 /
